refactor - 의존성 주입 final class 분리 및 프로토콜 분리 #78
