+++ SideDishProject/SideDishProject/SceneDelegate.swift
@@ -21,14 +21,4 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     }
 }
 
-final class ProductSceneContainer {
-    static func makeRoot(repository: ProductRepository) -> UIViewController {
-        let viewController = HomeViewController.create(with: repository)
-        return RootNavigationController(rootViewController: viewController)
-    }
-    
-    static func makeDetail(repository: ProductRepository) -> UIViewController {
-        return DetailViewController.create(with: repository)
-    }
-    
-}
+
